projectiles now return damage value correctly as float instead of int

+++ Game/Entities/Projectile.cpp
@@ -49,7 +49,7 @@ float Projectile::getMaxSpeed() const
     return mSpeed;
 }
 
-int Projectile::getDamage() const
+float Projectile::getDamage() const
 {
     return table[mType].damage;
 }
